Greatly simplify the task scheduler
It previously had a lot of logic to attempt to provide thread-safety
by maintaining a collection of queues.  Of course - duh - JS is
single-threaded by design and so that wasn't needed at all.

Also, there is no benefit in keeping a reference to any promises
which are used to trigger the execution of the queue.  We don't
offer any functionality to abort the queue.